refactor : Service 추상화
